chore: render paymentsmenu component per route
